+++ library/core/src/main/java/com/google/android/exoplayer2/util/EventLogger.java
@@ -27,7 +27,7 @@ import com.google.android.exoplayer2.RendererCapabilities;
 import com.google.android.exoplayer2.Timeline;
 import com.google.android.exoplayer2.audio.AudioRendererEventListener;
 import com.google.android.exoplayer2.decoder.DecoderCounters;
-import com.google.android.exoplayer2.drm.DefaultDrmSessionManager;
+import com.google.android.exoplayer2.drm.DefaultDrmSessionEventListener;
 import com.google.android.exoplayer2.metadata.Metadata;
 import com.google.android.exoplayer2.metadata.MetadataOutput;
 import com.google.android.exoplayer2.metadata.emsg.EventMessage;
@@ -60,7 +60,7 @@ public class EventLogger
         VideoRendererEventListener,
         MediaSourceEventListener,
         AdsMediaSource.EventListener,
-        DefaultDrmSessionManager.EventListener {
+        DefaultDrmSessionEventListener {
 
   private static final String TAG = "EventLogger";
   private static final int MAX_TIMELINE_ITEM_LINES = 3;